Rename package hamcrestext to hamcrestopt.
The name hamcrestext is outdated because the project'S name changed from
hamcrestext to Hamcrest Optional.

Fixes #18.
